backend/requirements.txt
requests==2.32.3
Django==5.1.2
djangorestframework==3.15.2
gunicorn==20.1.0
channels==4.2.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
h11==0.14.0
httptools==0.6.4
idna==3.10
python-dotenv==1.0.1
pyyaml==6.0.2
sniffio==1.3.1
uvicorn==0.34.0
watchfiles==1.0.4
daphne==4.1.2
requests==2.32.3